Re-design tokenizer
